requirements.txt
asgiref==3.8.1
Django==4.2.19
django-cors-headers==4.7.0
django-easy-audit==1.3.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
mssql-django==1.5
mysqlclient==2.2.7
numpy==2.0.2
packaging==24.2
pandas==2.2.3
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
